requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.5.2
astroid==2.8.5
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.7
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.1.4
django-cors-headers==3.10.0
django-crontab==0.7.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
geographiclib==1.52
geopy==2.2.0
google==3.0.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
hiredis==2.0.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==1.0.4
msgpack-python==0.5.6
password-strength==0.0.3.post2
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.11.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
redis==2.10.6
requests==2.26.0
rsa==4.7.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
twilio==6.56.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==0.57.0
wrapt==1.13.3
zope.interface==5.4.0
